# evm: make profiler more accurate