--- a/packages/evm/src/evm.ts
+++ b/packages/evm/src/evm.ts
@@ -278,11 +278,15 @@ export class EVM implements EVMInterface {
     let result: ExecResult
     if (message.isCompiled) {
       let timer: Timer
+      let callTimer: Timer | undefined
       let target: string
       if (this._optsCached.profiler?.enabled === true) {
         target = bytesToUnprefixedHex(message.codeAddress.bytes)
         // TODO: map target precompile not to address, but to a name
         target = getPrecompileName(target) ?? target.slice(20)
+        if (this.performanceLogger.hasTimer()) {
+          callTimer = this.performanceLogger.pauseTimer()
+        }
         timer = this.performanceLogger.startTimer(target)
       }
       result = await this.runPrecompile(
@@ -293,6 +297,9 @@ export class EVM implements EVMInterface {
 
       if (this._optsCached.profiler?.enabled === true) {
         this.performanceLogger.stopTimer(timer!, Number(result.executionGasUsed), 'precompiles')
+        if (callTimer !== undefined) {
+          this.performanceLogger.unpauseTimer(callTimer)
+        }
       }
       result.gasRefund = message.gasRefund
     } else {