@jochem-brouwer 2024-01-20
Ok, now also added timers for the initialization and the overhead. However, the reported opcode time is now longer than the reported time that `evm.runCall` should use :thinking: I'm not sure because the Logger has explicit logic that only one timer can be running at a time :thinking: