@jochem-brouwer 2024-01-20
I should add the timers to precompiles too, `npm run test:state -- --profile --test=CALLBlake2f_MaxRounds` fails with profiler enabled.