@jochem-brouwer 2024-01-20
Tested on client and seems to work.